🚚 move get_user_id_by_user_email to user repository
