+++ backend/modules/user/userIdentity/entity/user.py
@@ -1,0 +1,10 @@
+from typing import Optional
+from uuid import UUID
+
+from pydantic import BaseModel
+
+
+class UserIdentity(BaseModel):
+    id: UUID
+    email: Optional[str] = None
+    openai_api_key: Optional[str] = None
