--- backend/repository/user/__init__.py
@@ -1,1 +1,0 @@
-from .get_user_id_by_user_email import get_user_id_by_user_email
