+++ backend/routes/subscription_routes.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from middlewares.auth.auth_bearer import AuthBearer, get_current_user
 from models import BrainSubscription, PromptStatusEnum
 from modules.user.userIdentity.entity.user_identity import UserIdentity
+from modules.user.userIdentity.repository import get_user_id_by_user_email
 from pydantic import BaseModel
 from repository.brain import (
     create_brain_user,
@@ -21,7 +22,6 @@ from repository.brain_subscription import (
     resend_invitation_email,
 )
 from repository.prompt import delete_prompt_by_id, get_prompt_by_id
-from repository.user import get_user_id_by_user_email
 from routes.authorizations.brain_authorization import (
     RoleEnum,
     has_brain_authorization,
